+++ cmd/aws-sso/ecs_client_cmd.go
@@ -32,7 +32,9 @@ import (
 	"github.com/synfinatic/gotable"
 )
 
-type EcsListCmd struct{}
+type EcsListCmd struct {
+	Port int `kong:"help='TCP port of aws-sso ECS Server',env='AWS_SSO_ECS_PORT',default=4144"` // SEE ECS_PORT in ecs_cmd.go
+}
 
 type EcsLoadCmd struct {
 	// AWS Params
@@ -69,15 +71,7 @@ func (cc *EcsLoadCmd) Run(ctx *RunContext) error {
 }
 
 func (cc *EcsProfileCmd) Run(ctx *RunContext) error {
-	clientCert, err := ctx.Store.GetEcsSslCert()
-	if err != nil {
-		return err
-	}
-	bearerToken, err := ctx.Store.GetEcsBearerToken()
-	if err != nil {
-		return err
-	}
-	c := client.NewECSClient(ctx.Cli.Ecs.Profile.Port, bearerToken, clientCert)
+	c := newClient(ctx.Cli.Ecs.Profile.Port, ctx)
 
 	profile, err := c.GetProfile()
 	if err != nil {
@@ -95,21 +89,15 @@ func (cc *EcsProfileCmd) Run(ctx *RunContext) error {
 }
 
 func (cc *EcsUnloadCmd) Run(ctx *RunContext) error {
-	clientCert, err := ctx.Store.GetEcsSslCert()
-	if err != nil {
-		return err
-	}
-	bearerToken, err := ctx.Store.GetEcsBearerToken()
-	if err != nil {
-		return err
-	}
-	c := client.NewECSClient(ctx.Cli.Ecs.Unload.Port, bearerToken, clientCert)
+	c := newClient(ctx.Cli.Ecs.Unload.Port, ctx)
 
 	return c.Delete(ctx.Cli.Ecs.Unload.Profile)
 }
 
 // Loads our AWS API creds into the ECS Server
 func ecsLoadCmd(ctx *RunContext, awssso *sso.AWSSSO, accountId int64, role string) error {
+	c := newClient(ctx.Cli.Ecs.Load.Port, ctx)
+
 	creds := GetRoleCredentials(ctx, awssso, accountId, role)
 
 	cache := ctx.Settings.Cache.GetSSO() // ctx.Settings.Cache.Refresh(awssso, ssoConfig, ctx.Cli.SSO)
@@ -130,31 +118,12 @@ func ecsLoadCmd(ctx *RunContext, awssso *sso.AWSSSO, accountId int64, role strin
 		log.WithError(err).Warnf("Unable to update cache")
 	}
 
-	// do something
-	clientCert, err := ctx.Store.GetEcsSslCert()
-	if err != nil {
-		return err
-	}
-	bearerToken, err := ctx.Store.GetEcsBearerToken()
-	if err != nil {
-		return err
-	}
-	c := client.NewECSClient(ctx.Cli.Ecs.Load.Port, bearerToken, clientCert)
-
 	log.Debugf("%s", spew.Sdump(rFlat))
 	return c.SubmitCreds(creds, rFlat.Profile, ctx.Cli.Ecs.Load.Slotted)
 }
 
 func (cc *EcsListCmd) Run(ctx *RunContext) error {
-	clientCert, err := ctx.Store.GetEcsSslCert()
-	if err != nil {
-		return err
-	}
-	bearerToken, err := ctx.Store.GetEcsBearerToken()
-	if err != nil {
-		return err
-	}
-	c := client.NewECSClient(ctx.Cli.Ecs.Profile.Port, bearerToken, clientCert)
+	c := newClient(ctx.Cli.Ecs.Profile.Port, ctx)
 
 	profiles, err := c.ListProfiles()
 	if err != nil {
@@ -186,4 +155,16 @@ func listProfiles(profiles []ecs.ListProfilesResponse) error {
 	}
 
 	return err
+}
+
+func newClient(port int, ctx *RunContext) *client.ECSClient {
+	certChain, err := ctx.Store.GetEcsSslCert()
+	if err != nil {
+		log.Fatalf("Unable to get ECS SSL cert: %s", err)
+	}
+	bearerToken, err := ctx.Store.GetEcsBearerToken()
+	if err != nil {
+		log.Fatalf("Unable to get ECS bearer token: %s", err)
+	}
+	return client.NewECSClient(port, bearerToken, certChain)
 }